remark libary add-ons
async is added to getPostData function in the posts.js , along with two variables (processedContent, contentHtml) which will process the string portion of metadata files

contentHtml will be returned along with id and data

await is added to getPostData to enable usage of remark library in getStaticProps function in [id].js under posts folder.

A div element is added after postData.date to show the string portion of associated metadata file.